update app to use interpolation model
