+++ app.py
@@ -5,31 +5,46 @@ import os
 import pickle
 from PIL import Image
 from pathlib import Path
-from shutil import copy
+from shutil import copy, rmtree
 import pandas as pd
 import numpy as np
 from glob import glob
-
+from copy import deepcopy
+import math
+import time
 import torch
 from torchvision import transforms as T
 from einops import rearrange
 from omegaconf import OmegaConf
 from ldm.data.generate_utils import InferenceModel, draw_styles, convert_fname, interp_mask
 
-device = 'cuda:0'
+DEVICE = 'cuda:1'
+CONFIG_FILE = 'models/upgpt/interp_256/config.yaml'
+CKPT = 'models/upgpt/interp_256/upgpt.interp256.v1.ckpt'
+upscale_ckpt = "models/upgpt/upscale/upgpt.upscale.v1.ckpt"
+
 styles_root = Path('styles')
 cache_root = Path('app_cache')
 local_style_root = cache_root/'styles'
 local_pose_root = cache_root/'pose'
 local_lowres_root = cache_root/'samples_lowres'
+local_interp_root = cache_root/'interp'
 
 os.makedirs(local_style_root, exist_ok=True)
 os.makedirs(local_lowres_root, exist_ok=True)
+os.makedirs(local_interp_root, exist_ok=True)
 pose_folders = sorted([x[0] for x in os.walk(local_pose_root)][1:])
 pose_images = []
 for pose_folder in pose_folders:
     pose_images.append(Image.open(glob(os.path.join(pose_folder,'*.jpg'))[0]))
 #pose_images = [Image.open(Path(x)/'pose.jpg') for x in pose_folders]
+
+def delete_files_in_folder(folder_path):
+    for file_name in os.listdir(folder_path):
+        file_path = os.path.join(folder_path, file_name)
+        if os.path.isfile(file_path):
+            os.remove(file_path)
+
 
 def clear_image_cache():
     for x in glob(str(local_lowres_root/'*')):
@@ -46,13 +61,14 @@ def get_image_number():
     
     return '{:03d}'.format(file_id + 1)
 
-def get_samples():
-    return [Image.open(x) for x in sorted(glob(str(local_lowres_root/'*.jpg')))]
+def get_samples(folder):
+    print(folder, sorted(glob(str(folder/'*.jpg'))))
+    return [Image.open(x) for x in sorted(glob(str(folder/'*.jpg')))]
 
 map_df = pd.read_csv("data/deepfashion/deepfashion_map.csv")
 map_df.set_index('image', inplace=True)
 
-st.title('UPGPT - mix-and-match text and visual prompts for human image generation')
+st.title('UPGPT - Person Image Generation, Edit, Pose Transfer and Pose Interpolation')
 style_names = ['face', 'hair', 'headwear', 'background', 'top', 'outer', 'bottom', 'shoes', 'accesories']
 
 
@@ -81,24 +97,20 @@ lr_transform = T.Compose([
     T.Lambda(lambda x: x * 2. - 1.,)])
 
 @st.cache_resource
-def upgpt_model(config_file = 'models/upgpt/pt_256/config.yaml',
-                ckpt = 'models/upgpt/pt_256/upgpt.pt256.v1.ckpt', 
-                device = 'cuda:0'):    
-    
+def upgpt_model(config_file, ckpt, device):    
     config = OmegaConf.load(config_file)
     model = InferenceModel(config, ckpt, device)
 
     return model 
 
-model = upgpt_model(device=device)
+model = upgpt_model(CONFIG_FILE, CKPT, DEVICE)
 
 disable_upscale = True
-upscale_ckpt = "models/upgpt/upscale/upgpt.upscale.v1.ckpt"
 if os.path.exists(upscale_ckpt):
     upscale_model = upgpt_model('models/upgpt/upscale/config.yaml',
-                                upscale_ckpt, 
-                                device)
+                                upscale_ckpt, device=DEVICE)
     disable_upscale = False
+
 def load_smpl(folder):
     smpl_file = glob(str(Path(folder)/'*.p'))[0]
     smpl_image_file = glob(str(Path(folder)/'*.jpg'))[0]
@@ -142,28 +154,84 @@ def get_styles(input_style_names=style_names):
     style_images = torch.stack(style_images)  
     return style_images
 
+def get_coord(batch_mask):
+    mask = batch_mask[0].cpu().numpy()
+    mask[mask==-1] = 0
+    x = np.nonzero(np.mean(mask,1))[0]
+    xmin, xmax = x[0], x[-1]
+    y = np.nonzero(np.mean(mask,0))[0]
+    ymin, ymax = y[0], y[-1]
 
-left_column, mid_column, right_column, sr_column = st.columns([1,1,1,1])
+    return np.array([xmin, xmax, ymin, ymax])
+
+def get_mask(mask, coord):
+    xmin, xmax, ymin, ymax = coord
+    new_mask = np.ones_like(mask.cpu().numpy())*(-1)
+    new_mask[0,xmin:xmax+1, ymin:ymax+1] = -0.99215686
+    #return new_mask
+    return torch.tensor(new_mask).to(mask.device)
+
+def interp_mask(src_mask, dst_mask, alpha):    
+    coord_1 = get_coord(src_mask)
+    coord_2 = get_coord(dst_mask)
+
+    coord = (alpha * coord_1 + (1 - alpha) * coord_2).astype(np.int32)
+    #print(coord)
+    #coord = np.array([ 0, 31,  12, 19])
+    new_mask = get_mask(src_mask, coord)
+    return new_mask
+
+
+
+left_column, mid_column, right_column = st.columns([1,1,3])
+
+# right column
 right_column.markdown("##### Generated Images")
 gen_image = right_column.empty()
-low_res_images = get_samples()
-image_ids = [i+1 for i in range(len(low_res_images))]     
 
-
-def display_samples():
+def display_samples(folder, loc=gen_image):
     global image_ids
-    low_res_images = get_samples()
+    low_res_images = get_samples(folder)
     image_ids = [i+1 for i in range(len(low_res_images))]
-    gen_image.image(low_res_images, width=192, caption=image_ids)
+    loc.image(low_res_images, width=192, caption=image_ids)
 
-display_samples()
+with right_column:
+    
+    show_image_button = right_column.button(label='Show images')
 
+    if show_image_button:
+        display_samples(local_lowres_root)
+
+    image_files = sorted(glob(str(local_lowres_root/'*.jpg')))
+    delete_ids = [i+1 for i in range(len(image_files))]
+    
+    del_options = st.multiselect('Select images to delete', delete_ids, [])
+    clear_image_button = st.button(label='Delete images')    
+    if clear_image_button:
+        for del_option in del_options:
+            os.remove(image_files[del_option-1])
+        #clear_image_cache()
+        #gen_image.empty()
+    delete_all_gen_button = st.button(label='Delete all generated images')
+    if delete_all_gen_button:
+        delete_files_in_folder(str(local_lowres_root    ))        
+    display_samples(local_lowres_root)
+
+right_column.markdown("##### Pose Interpolation")
+interp_factors = right_column.text_input('Interplation factor, from 1.0 to 0.0, comma seperated. You may need to tweak the spacing for better result.', 
+                            value='1.0, 0.8, 0.7, 0.6, 0.4, 0.3, 0.2, 0.1, 0.0')
+interp_image = right_column.empty()
+delete_all_interp_button = right_column.button(label='Delete all images')
+if delete_all_interp_button:
+    delete_files_in_folder(str(local_interp_root))
+display_samples(local_lowres_root)
+display_samples(local_interp_root, interp_image)
 
 
 with left_column:
     with st.form(key='input'):    
-        st.markdown("##### Text Prompt")
-        default_text = "a woman is wearing a sleeveless tank and a short skirt."
+        st.markdown("##### Content Text")
+        default_text = "a woman is wearing a long sleeve shirt and long pant."
         content_text = st.text_area('Content text', label_visibility='hidden', value=default_text)
         st.markdown("##### Style Text")
         
@@ -177,10 +245,19 @@ with left_column:
         st.markdown("##### Pose")
         pose_ids = [i+1 for i in range(len(pose_images))]
         st.image(pose_images, caption=pose_ids, width=96)
-        pose_select = st.radio("Select pose", pose_ids)
+        pose_column_1, pose_column_2 = st.columns([1,1])
+        with pose_column_1:
+            pose_select = st.radio("Source pose", pose_ids, index=0)
+        with pose_column_2:
+            target_pose_select = st.radio("Target pose", pose_ids, index=3)
         st.markdown("---")
         
-        submit_button = st.form_submit_button(label='Generate')
+        gen_column, interp_column = st.columns([1,1])
+        with gen_column:
+            submit_button = st.form_submit_button(label='Generate')
+        with interp_column:
+            interp_button = st.form_submit_button(label='Interpolate')
+
         if submit_button:
             style_features = get_styles()
             batch = {}
@@ -194,17 +271,48 @@ with left_column:
             log = model.generate(batch, 200)
             sample = Image.fromarray(np.uint8(log['samples'][0]*255))
             sample.save(local_lowres_root/f'sample_{get_image_number()}.jpg')
-            display_samples()
+            display_samples(local_lowres_root)
             # gen_image.image(get_samples(), width=192)
             # low_res_images = get_samples()
             # image_ids = [i+1 for i in range(len(low_res_images))]
+
+        if interp_button:
+            delete_files_in_folder(str(local_interp_root))
+            style_features = get_styles()
+            batch = {}
+            style_texts_dict = dict(zip(style_names, style_texts))
+            batch['image'] = image_transform(Image.open(cache_root/'image_256.jpg')) # dummy
+            batch['styles'] = model.mix_style(style_features, style_texts_dict)
+            batch['txt'] = content_text
+            
+            dst_batch = deepcopy(batch)
+            src_pose = load_smpl(pose_folders[pose_select - 1])
+            dst_pose = load_smpl(pose_folders[target_pose_select - 1])
+            batch.update(src_pose)
+            dst_batch.update(dst_pose)
+
+            alphas = np.array([float(num) for num in interp_factors.split(',')])
+            batch = model.create_batch(batch, repeat=len(alphas))
+            
+            for i, alpha in enumerate(alphas):
+                batch['smpl'][i] = alpha * batch['smpl'][i] + (1 - alpha) * dst_batch['smpl'].to(DEVICE)
+                batch['person_mask'][i] = interp_mask(batch['person_mask'][i], dst_batch['person_mask'], alpha)
+            log = model.generate(batch, 200)
+            
+            for i, sample in enumerate(log['samples']):
+                sample = Image.fromarray(np.uint8(sample*255))
+                sample.save(local_interp_root/f'interp_{i}.jpg')
+                #interp_images.append(sample)
+            time.sleep(1) # wait to save file into disk
+            display_samples(local_interp_root, interp_image)
+            #gen_image.image(interp_images, width=192)
 
 with mid_column:
     #left_2_column, right_2_column = st.columns([1,1])
     #style_image = right_2_column.empty()
     #with left_2_column:
     with st.form("my-form", clear_on_submit=False):
-        st.markdown("##### Image Styles")
+        st.markdown("##### Style Images")
         style_file = st.file_uploader("Style reference")
         style_image = st.empty()
         options = None
@@ -249,30 +357,35 @@ with mid_column:
 
                 
 
-with right_column:
+# with right_column:
     
-    show_image_button = right_column.button(label='Show images')
+#     show_image_button = right_column.button(label='Show images')
 
-    if show_image_button:
-        display_samples()
+#     if show_image_button:
+#         display_samples(local_lowres_root)
 
-    image_files = sorted(glob(str(local_lowres_root/'*.jpg')))
-    delete_ids = [i+1 for i in range(len(image_files))]
-    del_options = st.multiselect('Select images to delete', delete_ids, [])
-    clear_image_button = st.button(label='Delete images')
-    if clear_image_button:
-        for del_option in del_options:
-            os.remove(image_files[del_option-1])
-        #clear_image_cache()
-        #gen_image.empty()
-    display_samples()
+#     image_files = sorted(glob(str(local_lowres_root/'*.jpg')))
+#     delete_ids = [i+1 for i in range(len(image_files))]
+#     del_options = st.multiselect('Select images to delete', delete_ids, [])
+#     clear_image_button = st.button(label='Delete images')
+#     if clear_image_button:
+#         for del_option in del_options:
+#             os.remove(image_files[del_option-1])
+#         #clear_image_cache()
+#         #gen_image.empty()
+#     display_samples(local_lowres_root)
 
-with sr_column:    
+with right_column:
     st.markdown('#####  Upscale')
-    upscale_select = st.selectbox('Upscale', image_ids, label_visibility='hidden')
-    clear_image_button = st.button(label='Upscale', disabled=disable_upscale)
-    if clear_image_button:
-        low_res_images = get_samples()
+    c1, c2 = st.columns([1,1])
+    with c1:
+        upscale_folder = st.selectbox('Generated/Interpolated', ['Generated','Interpolated'], label_visibility='hidden')
+    with c2:
+        upscale_select = st.selectbox('Upscale', image_ids, label_visibility='hidden')    
+    upscale_button = st.button(label='Upscale', disabled=disable_upscale)
+    if upscale_button:
+        folder = local_interp_root if upscale_folder == 'Interpolated' else local_lowres_root
+        low_res_images = get_samples(folder)
         style_features = get_styles(style_names)
         batch = {}
         batch['image'] = image_transform(Image.open(cache_root/'image_512.jpg')) # dummy
